+++ packages/stacked_tools/test/helpers/test_helpers.dart
@@ -114,8 +114,8 @@ void registerServices() {
 // @stacked-mock-helper-register
 }
 
-void createTestFile() {
-  File(ksTestFileName).writeAsStringSync(kAppTemplateAppContent);
+void createTestFile(String template) {
+  File(ksTestFileName).writeAsStringSync(template);
 }
 
 Future<void> deleteTestFile() async {